Move some resource normalization classes to a new package since the resource package was getting crowded
