+++ storm-server/pom.xml
@@ -130,7 +130,7 @@
                 <artifactId>maven-checkstyle-plugin</artifactId>
                 <!--Note - the version would be inherited-->
                 <configuration>
-                    <maxAllowedViolations>2620</maxAllowedViolations>
+                    <maxAllowedViolations>2617</maxAllowedViolations>
                 </configuration>
             </plugin>
             <plugin>
